refactor: delete thread

+++ app/api/thread/route.ts
@@ -50,9 +50,12 @@ export async function PUT(req: NextRequest) {
 }
 
 export async function DELETE(req: NextRequest) {
-  const data = (await req.json()) as Pick<Thread, 'id'>
+  const searchParams = req.nextUrl.searchParams
+  const id = searchParams.get('id')
+  if (!id) throw new Error('there is no id')
+
   const cate = await prisma.thread.delete({
-    where: { id: data.id },
+    where: { id },
     include: { Message: true }
   })
   return NextResponse.json(cate)
